Add/configure rubocop
By default rubocop is a tad overzealous.  We use the config from the
govuk-lint[1] repo to rein it in a bit and we expect to modify it to our
liking over time.

Rubocop version is pinned in Gemfile as recormended and binstubs are
generated.

We exclude the 'bin/*' dir since files generated there are generally out
of our control.

[1] https://github.com/alphagov/govuk-lint

fixup: add rubocop/gem/binstubs/exclude-bindir

fixup: add rubocop/gem/binstubs/exclude-bindir

rubocop binstub