Favour documentation over scripts/envrc magic
Previously we had some .envrc magic and some scripts for running the app
against a paas-cf dev environment and the README was geared towards
this.

We want to try to stay a little more agnostic about who might be running
this (even if it is just us for now).

So we will favour adding more generic instructions to the README along
with outling the scripts for how to get up and running quickly with
GOV.UK PaaS