Inherit rubocop config from govuk-lint
By inheriting rubocop configuration from the govuk-lint gem we can stay
in sync with upstream changes by updating the gem.

We extend the base config to enable the Rails and Rspec cops then rein
in the Rspec cops a bit as they are a bit too agressive by default.

The large-block cop is excluded from the rails config dir since it is
expected to always have large blocks.