webpack configuration for govuk_template
We want to use the govuk_template_jinja module without having to rely on
manually copying out templates, css and assets.

To do this we need to preprocess the templates and css a bit to ensure
that the paths of all src="" attributes in njk files and url()s in css
files refer to the correct places.

Unfortunatly the main template assumes that we would be altering the {{
asset_path }} at RUNTIME not COMPILE time so we have to fix this with a
regex replace.

We believe that the govuk-frontend project will replace the
govuk_template eventully so we don't expact to need this code forever