enable lint warning for use of console.log
We want all our log output to be structured and performant. console.log
is neither of these things, so we should avoid it and use the configured
logger.

We enable the "no-console" linting rule so to get warnings from the
linter about any use of console.log. We don't want it to be an "error"
since using console.log for "print debuging" is very common and it
would be annoying to have `npm test` fail (return non-zero exit) for
this.