output an "all tested pass" message
`npm run test` combines all the various tests (js linting, sass linting,
coverage, unit tests) and some of these tests may output warnings that
are not critical for a pass but look scarey (like 'fixme' comments for
example).

This outputs a comforting message as the last line if we are exiting
successfully to make it clear that everything is ok.