Use pino for structured logging
We want to emit structured logs, and we will use `pino` for this.

We create some express middleware to capture and emit all the HTTP request
data.

The express `req` object contains an instance of the logger with the
current request id attached to keep a trail of logs:

app.get((req) => {
	req.logger.info('hi')
})

...and a logger can also be fetched via:

import logger from './logger';