Create a strong content-security-policy
We want to ensure we send out a strong Content-Security-Policy[1]

We block all inline script tags and script from unknown sources.

Because the govuk_template template includes two inline script blocks we
have had to add the hashes of their content to "allow" them to execute.

We also allow google-analytics since we based this CSP on our product
page's policy and expect we will be adding analytics at some point.

[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP