Vendor @govuk-frontend package
We are depending on the govuk-fronend package to supply the css and
templates from the GOV.UK Design System[1]

Currently this package is in private beta and so not available from npm
to the general public. We have been given a username/password to access
updates to the module but this is not very practical for day-to-day use
so we have vendored the current version.

[1] https://govuk-design-system-production.cloudapps.digital/