Add blackbox/server testing
We want to be able to test the behaviour of the server process itself
(blackbox) as well as the existing handler-level unit testing.

To do this the bulk of the logic for starting/stopping the server and
how hot reloading of handlers works is broken out into a Server module
that can be tested, which allows us to keep main.js solely responsible
for setting up the server making it more suitable for testing.

main.test.js was created as a "blackbox" test that spawns the process
and checks that:

* it can be configured via environment variable PORT
* it responds gracefully to SIGINT / SIGTERM

To aid testing and reduce the chance of port clashes the default
behaviour of starting the server when no PORT is set is now to pick a
random port number.

An eslint overide was added for src/main.js to allow the use of
process.exit - since the job of main.js involves handling the exit code
this is reasonable