export app as a function to allow configuration
We want to be able to pass configuration to the main 'app' handler so
that we can set things up without relying on globals.

For example we want to be able to pass in configuration for the logger
from main.js or from tests, we will want to pass in configuration for
api keys each from main.js and maybe something different from tests.

This changes the app export to be an init function that accepts a config
object rather than the preconfigured app.

The global src/logger module is removed in favour of passing in your
favourite logger during app initialization function.