`npm run watch` incremental build/test runner
Running `npm test` or `npm test:unit` is quite slow because it executes
a full build.

This implements running the tests (ENABLE_TESTS=true) via an onBuildExit
hook so that they are run immeditely after an incremental build
(ENABLE_WATCH=true).

An `npm run watch` script is added to easily watch the tests.

`ENABLE_TESTS=true npm start` would enable running the tests at the same
time as the hot-reloading server, but the output would be a bit messy so
probably best done seperately.

A new flag (ENABLE_COVERAGE=true) is added to control istambul/nyc
coverage so that the instrementation step can be skipped when not
needed.