Enable gzip compression
We should be serving responses with Content-Encoding: gzip to reduce
filesizes where possible.

All static assets are gziped at compile time to remove the need to compress
them on-the-fly during serving them.

All dynamic app responses are compressed on-the-fly. This can be
disabled by setting a `Cache-Control: no-transform` header on the
response.

As part of this change there was a need to process the assets discovered
from the govuk_template css (which were previously being emitted to the
asset dir with the help of the postcss-url helper). This is no longer
required and we can let the main file-loader handle the asset
processing to take advatange of the gzip.