Add oauth2 middleware for UAA integration
We implement a new middleware that provides all the required
functionality to integrate the app with UAA and provide authentication.

This middleware relies on passport and passport-oauth2 middlewares.

It implements oauth2 authentication by authentication codes, by
providing:

 - a /auth/login endpoint, which would redirect to the IDP authorize
endpoint for authentication.
 - a /auth/login/callback, to redirect when the authentication is
completed. Passport will then call the IDP token endpoint to retrieve
the token
 - a /auth/logout, which logouts using the passport functions.
 - a global handler that checks if the user is authenticated.

   Note: Using a global handler means that every request would
   be authenticated after this point, including missing endpoints (404)

This middleware is configured by passing a `config` object with the IDP
endpoints and client credentials, and the serverRootURL, required for
the callback.

We request the following scopes from UAA: cloud_controller.read, openid,
and uaa.user. But this functionality must be tested in integration
tests.

In the test we simulate the interaction of the user during the Oauth2
process, and we mock the response of the IDP token endpoint using the
nodejs nock library.