Integrate app.js with auth middleware
We add the auth middleware to our app.js.  The auth middleware is
mounted after all the public endpoints (/, assets, etc.), meanwhile any
private must be mounted after.

This approach has the collateral effect that we do not respond 404 to
unknown paths for non authenticated sessions, returning a 302 to
/auth/login instead. As this is a admin page, we do not consider such an
issue at the moment. This is open to discussion.

We read all the required configuration from environment variables in the
main.js, passing it into the config object.

The app.js would configure session persistance using cookies, so that
the session (including the UAA jwt token) would be stored in the cookie.

We did not implement encryption of the cookie itself, as are not storing
any critical information and we assume that the service would use HTTPS,
but we enable cookie signature (not tested yet).

In this commit we also fix some issues in the main.test.js when spawing
the nodejs process to test.