Upgrade to webpack 4.0
This update webpack to use the offical 4.0 release.

Webpack 4 came out of beta. We building with the beta version since our
work on this project was happening right on the cusp of the 4.0 release
which had some very appealing features such as less configuration
requirements.

We were previously working around an issue with the beta version's
"optimization" configuration where there were some inconsistancies with
chunks combined with target:node which cause build issues in production
mode... this is now resolved and we can set
runtimeChunk:false and allChunks: false to disable automatic bundle
chunks (chunks are of little use to a target:node build).

This allowed us to clean up how we were generating the test bundles in
NODE_ENV=production too ... previously we had to exclude all
node_modules to build the tests due to the issue with the chunking
described above, which would not have been giving us a 100% accurate
representation of the prod builds... now the tests are built in the
exact same way with only the test frameworks excluded