Refactor tests not to use supertest.expect()
With the current use of `supertest` we can experience some unclear
issues which result in hanging the tests without any notification as to
what went wrong. For instance, the following code:

```js
request(app)
  .get('/')
  .expect(nonExsistingVariable);
```

Should in fact throw an exception and stop the tests. The reality, is
that the exceptions are being caught and could then be handled with
`.catch()` method.

We thought it's unreasonable for us to add that after every `.except()`,
therefore we decided to go with different approach.

It proved difficult to replace `supertest` with another solution, due to
the goodness it provides when setting up and tearing down the
application on the go. We decided to refactor it in a way to get the
best of both worlds. Keep the `supertest` magic and not use the bugged
`.except()` function.

We're forcing the request to finish, by prefixing it with the `await`
key word and then taking the response to be used with the default tap
functionality.

!!! This does not get rid of the issue... It should however discourage
people from using or knowing of the faulty functionality.