Add organisation templates
Add template for showing organisations using the GOV.UK design system.
Use design system for title and fonts.

We may want a path for having no organisations and will probably want to
look at this later.

First pass at creating new component for navigation panel on this page
as it does not exist in the design system.  As this is specific to our
app we will keep it here for now.  Roughly following the pattern of the
design system for this component.  However some more work needs to be
done on it to ensure that it's accessible and that the structure of it
is reasonable.  It also needs a readme to describe more thoroughly what
it is.