Implement CF API client
The ones available for grabs, are out of date and could go distinct.
We're going to create our own mini client, that will do all the
necessary curl requests for us and return simple values.

We're adding `axios`[1] to our dependencies in order to make some
requests across the web. We could be using the `http`[2] but we'd like
to have an ability of promises.

[1]: https://github.com/axios/axios
[2]: https://nodejs.org/api/http.html