Make the Auth middleware check for expired token
It turns out Passport has no idea we're dealing with JWT, meaning we
have to handle checking the expiration ourselves.

We're adding `jsonwebtoken` as a dependency for quick and easy decoding
of a token on a fly. It doesn't unfortunately allow us to validate the
token as it requires a secret to be fully verified.

Instead, we're decoding the Token and manually comparing the timestamp
with the current one to see if we've expired or good to go.

If the token has expired, we're going to kick the user back to the login
page for re-authentication.