Remove returnUrl functionality
This is a nice feature allowing us to be redirected back to the place
we've left before the application decided we're supposed to be
authorised to view that page.

As convenient as it is, it is a potential low level security threat. It
could be used to manipulate users to perform an action they didn't
intend to. For instance, as an actor I could make a link to redirect to
app deletion final URL, and a not logged in user would be redirected to
the login page and then back to finish `authorised` action.

This may be unlikely scenario, but catching this threat early may save
us later in the future.