Establish a syncHandler functionality
We cannot justify the use of synchronous approach express enforces. It
may be confusing to some people on the team as to why and when we use
async/await or promises.

We're creating a simple handler, that will take care of that for us.
It's a matter of including it when needed. Refactor to come later.