Introduce UAA client
We'll need some functionality around inviting the user to the platform.
This is done through UAA, meaning we should have a separate API setup.

We also have an ability to find the user to check whether they already
exist or not.

This client makes additional call with the use of ClientID and
ClientSecret in a form of Basic Auth in order to retrieve the access
token to use for further calls.