Implement role checker
In various places of the application we'd like to be checking if the
user has the right permissions to view or manage the content. We're
providing a wrapper function to deal with it as well as the middleware
for easier use in blocking routes.

The trick is to also allow administrators the same privlidge due to
their status.