Make use of singing key
We'd like to verify the token user holds in order to check if they're
admin user when we need to. It means there needs to be a call to obtain
the public key from UAA. We can refactor our current use of `jwt.decode`
to do `jwt.verify`, meaning we can use less code to achieve the same
thing, rely on `jwt` library to perform validation and retrive the
payload when required.

We're going to load that token when we need it first and then will
attempt to store it whithin the UAA library itself.