Update README
We'd like to make it clearer which version is required to run the
application. The current wording is not helpful a lot to some people who
might not have the experiences with the use of node packages. For
instance, the Node LTS version is `8.9.0` but it's named `8 LTS`.

We're also removing some tabs in favour of spaces.