Remove the 'home' module
This has been a starting point for the application to get moving around.
As the application grew, we no longer need it to exist nor do we want it
to be presented to our users.

We took the approach of redirecting the root URL to `/organisations`
endpoint, as this would be ideally a starting point for the user.

The redirection is and should be a temporary solution, for we need to
implement the router system that would fix the routes in general.