Add sub navigation to all pages that deserve it
Previously, we had to include the sub navigation on every page that
might have required it. The approach quickly proven we will forget about
certain places meaning the sub navigation will not be displayed.

We're now including the if statement in the main layout, checking
whether an `organization` variable has been provided from the controller
and decide on displaying the nav at all times.