Add logo
Add logo for Platform as a Service.  This is in a separate div from the GOV.UK logo which isn't ideal as they both point to the same place.  To put them together we would have to take a copy of the template and modify it rather than simply feeding it in from the node-module which would mean that we could miss updates to this module.

Hopefully this template will be deprecated soon and superseded by a new version from the design system, so it's probably better to wait until this new version comes out.  If the two logos causes confusion for users before the new design system version comes out, we may need to revisit this however.