Introduce TypeScript
After finishing the initial MVP version of paas-admin, we believe going
with TypeScript would save us a lot of time in attempting to debug
certain parts of the application.

Providing types will assure we're using the functionality correctly and
will error in compilation time when we're not tying the knots.

It may be easier for other members of the team to find their selves
around the codebase as we're all use to and somewhat liking Golang.

Along with TypeScript, we're introducing linters to insist on certain
code styling.