Change the default logging level to warning
As we run the appliction now, it is becoming more annoying to look at
the noise the logs are generating. The access logs are gathered by the
user, meaning we can switch the node access logging off entirely.

It also could have been a security concern as session cookies have been
logged out...