Only amend permissions on change
We were making a lot of calls each time a single permision would change.
Meaning we have likely hit the problem with API responging with errors
when "unsetting" permission that has not been previously set.

We decided to record current and desired permission in HTML for later
comparision in the handlers and preparing the call that matters to end
user.