Less verbose logs
Pino by default would provide a lot of indormation through to the logs.
This may be helpful at some point, but in practice is annoying and
dengerous as a session cookie is printed.

We have a habbit of copying the error log and pasting in places, meaning
it could potentially leak the session. Let's avoid it by serializing the
request and response for the logger to use.