Fix jumping on space hover
Currently, when one would hover over the space box, all of the boxes
below would jump down by 2px. This is not really nice to look at or
experience and the solution may be relatively simple.

We're reducing the padding inside the box on hover, forcing the box size
to remain the same. This eliminates the issue.