Enforce date range to be provided in statement url
We want to force the statement page to only show for a month period so
we only accept the rangeStart param and then make the range be +1month

Install moment.js

We're handling dates, therefore we'd like to use a library that we trust
does it correctly. Moment is that library.