Simplify the usage of the token
We'd like to be sharing the token throughout the application. This
involves verifying the token against the signing key. We decided to
refactor our current usage to satisfy the need of sharing a single user
that has an ability to check if the desired scope is present.

In order to do that, we had to go back and readapt how it's being
currently used throughout `users` module.

It's also a refactor towards fixing multiple signingKeys issue

We have a story in the backlog to fix the issue some people in the team
have encountered. This isn't really solving the issue, but makes it
easier for us to implement the fix in the near future.