We cannot rely on pinning MINOR engine version
We previously pinned the node js engine to the MINOR version number as
we believed it to be rare that MINOR versions were removed from
buildpacks while PATCH versions were removed frequently.

We have seen that the nodejs buildpack periodically removes both MINOR
and PATCH versions so to avoid buildpack updates breaking our
deployments we will only pin to the MAJOR version and rely on the
buildpack to ensure a suitably up to date version is used from the
LTS/Carbon 8.x.x line is used.