Use /v2/info to obtain UAA endpoints
Currently, we have to provide a lot of environment variables before
running the application. This is annoying some of these could be sliced
out.

We've stumbled upon the misconfiguration issue, where the Google OAuth
is set with redirection_uri pointing at `login.*`. Which furthermore
prevents admin users from using the pazmin tool.

In reality these two (`uaa.*` and `login.*`) are pointing to the same
application. We didn't feel comfortable setting it to `uaa.*` in Google
configuration as we may choose to have a different auth application in
the future.

Instead, we're going to request `/v2/info` and set those up in the
application itself, to reduce manual configuration.