Implement pricing calculator
We're delivering a functionality to allow our users to generate what
pricing they'd be facing whilst choosing GOV.UK PaaS.

The billing API would fallout if the GUID would not be unique, therefore
we're introducing `uuid` library and its types to make it easier for
ourselves to interact with API.

Unfortunately, the new route needs to be outside the AUTH gate, meaning
for the time being it doesn't fit our cool and shiny router and we have
to make it part of express routing instead.

We're fixing IParamters, as it no longer appears that the data received
is always going to be string to string map...

Adding a method to tap typings to satisfy our linter and compiler.