Requiring acceptance of terms documents on login
We need to present terms to uses who have not yet signed them.

We have created an AccountClient to interface with the paas-accounts
API, and middleware that uses it to check the current user has accepted the
current terms documents and if not redirects them to a page requesting
they "sign" the document.

The terms content is in markdown format and is rendered with the help of
a markdown nunjucks filters built on showdown.