Add a markdown filter for nunjucks
We are using markdown format for the TermsOfUse documents. To enable
rendering the markdown content in the page a new "markdown" filter has
been added to the global nunjucks template environment that allows you
to pipe strings to it an get back html.

Example:

```
{{ mycontent | markdown | safe }}
```

The 'showdown' package was added as a dependency that does the
conversion.