Remove use of special linkTo function in layout
We need to use the layout in places that may not have the router's
helper functions like linkTo available.

The layout was only using the linkTo helper to point to the home page,
which was easy to replace with a "/" for now.

This also means that the errors templates do not need to load the router
which was not ideal