Swap out snyk for npm audit
We have decided to use an `npm audit` based check for checking
vulenerabilies in dependecies as part of the `npm test` script.

`npm audit` is still quite young and as such does not have a way to only
return error codes for high/critical issues or a way to whitelist known
problems that we have decided are not going to cause problems (such as
things that are only used in dev or issues in code paths we never
touch).

To work around this limitation we have added a simple grep based script
to only bail when a high/critical vulernability is seen. Once `npm
audit` has added more features (see github issues) we can improve this.

This change removes the dependecy on snyk which was providing similar
function but had been disabled due to false positives.