Update govuk-frontend paths for macros/scss
The latest version of the govuk-frontend stuff move things around this
updates the paths to their new locations.

We wanted to keep the longer/full path names to the macros so it was
clear where the macro was being loaded from, however internally the
govuk-frontend macros/templates use shorted names like "button/macro"
and so we needed to update the includePaths library path in webpack to
allow these shorter paths to continue working.