Make the Calculator Summary Sticky
Make the Calculator Summary Sticky but using sticky CSS from HTML5.  This is not supported by older browsers.  However we tested this in browserstack with various versions of Internet Explorer and when it doesn't work it just reverts back to being a fixed element.  A sticky effect provides a nice UI but is not vital for using the calculator.