Add Spacing and Tone Down Button
Add whitespace and make buttons look like links.  We don't usually make buttons look like links; however the style we have for buttons is very bright and designed to bring attention to it as a single element on the page.  In this case we have numerous buttons on the page and the usual style is too loud.  The new style is more subtle but is still clearly an action that users can perform.