Add Copy and Comments
Add headings for service list table to make it clearer what the select boxes allow you to chosse a service to add.  Also add comment explaining that we are having to hide the labels with CSS so they are not visible to anyone as the Gov.UK Design System macros currently include HTML which should be optional.