Enable overview of CUPS
There is a little API inconvenience, where the call listing all the
services in the space, will return mixed results for both user and
system provided instances. The endpoint however, to summarise each is
not working in a mixed manner, meaning we need to make a different call
for each of these.

API being inconvenient, there is no way to differentiate these services,
therefore we need to make additional call, to get all CUPS only, and
then doing a quick loop through them to see if the requested GUID
matches one of them. Based on that, we're making a targeted call to the
API.