Allow user deletion
We'd like to allow our tenants to remove certain users from their
orgranisation. This can be done via the secret `users` role being unset
on already existing org user. We're adding views for delete
confirmation and the success page.

We need to amend the CF client in order to force the resource to be
deleted recursively.