Reimplement how state is updated + "app" resources
Previously we had both the current state and an operation (like
"remove" with extra args) in the query string, which then calculated the
final state. This meant there were multiple urls to the same state and
had a fairly complex method of calculating the state.

This has been replaced with an implementation that generates hidden form
fields for the current state and so each submission submits the state.

Some of the (unwanted) code around splitting out the serviceName and
version from the plan names has been improved a bit and a system for
whitelisting only the services that we are happy showing has been added.
Hopefully this code can be cleaned up as we continue to improve
paas-billing's responses around plans.

We were also missing the fields for the "app" resources (numberOfNodes
and memoryInMB). These have been added into the form.