Fix sticky and remove stray debug
The sticky effect on the summary box for the calculator became broken somewhere along the way and so this is fixed.  It currently isn't visible as the list of services isn't currently long enough for sticky to happen, but if the list of services does become longer this design won't be lost.

Also removed a stray parameter dump as this would have been for debugging purposes.