Add resend invite functionality
We'd like to allow our tenants to resend user invites, when they know
a link has expired or a user has been unable to log into the platform.

Unfortunately, the `active` flag on CF has not been updated when the
user finishes up setting their account, meaning we have to display it at
all times. That's not a bug, it's a feature - Org manager can reset
user's password :D