Fix lack of default permission values
We've introduced a bug in the df91f55, where we got rid of the default
values in favour of 1/0 setup. This was a good decision and fixed
another issue, however causes the invite page, to not set the default
values, due to missing logic.

That has an reaction of attempting to delete roles from non-existing
user causing the application to 500 and not setting the desired
permissions.

It also had a bug where the state would be set to all boxes ticket,
which on save would have the user granted all permissions.