Fix how the space quota is displayed
When displaying the space quota in the list of spaces we
check if there is a space quota to show the space limit. But we
incorrectly refer to the wrong valiable, as the quota attribute
of the space has a `entity` field.

This causes the interface to show a NaNgb from a null/1024 operation.

We update the interface to use the right value.