Change github PR template markdown syntax
When doing a PR with [`hub`](https://github.com/github/hub), any
line starting with # will be ignored and not added to the PR message.

Our template uses # for the titles in the markdown syntax, so doing
a PR with `hub` would omit those lines. Using the alternate syntax
with undelying ---- fixes the issue.