Add permissions flag to templates
We'll need to pass around few variables, in order to display different
elements to the end user from our templating system.

We're duplicating a lot of work here, which I'm not particulary proud
of, but it'll only show us the need of sorting it out in a different
less painful maner.