Fix display user spaces membership
We were retrieving the users space membership but not displaying
it in the list of users. We need to asign this to some variable
and pass it to the view.

This still fails trying to retrieve the membership of other team
members when the user is non a manager, but at least we show
it for the org managers.

We also fix the link in the view.