Add functionality to statements
We'd like to make the statements more appealing to our tenants. We're
modifying some HTML layout, SCSS styling and backend to provide us with
the data we need.

We should have a dropdown with a last years worth of periods for
statements that will redirect to the correct statement on form
submission.

We should display the no data message, if the period doesn't have any
billing data - for instance before the organisation existed.

We should display the month in progress message, if the statement isn't
complete yet and is likely to change.