Change team members explanation section
I've replaced "superuser" with user, but I'm not clear if there was a reason superuser was here as it's different from any of the suggested text in any iteration.  The org manager is a superuser compared to the other users, but the distinction in the level of power they have is also explained in their description.

I'm taking "tasks" from the prototype copy, rather than "actions" from Paola's spreadsheet as I vaguely remember having a conversation about this and I think we decided that "tasks" made more sense than "actions" in the context of spaces.