Add more information about user level permissions.
More information about what users can do depending on what level they are.  We may need to check how much sense this makes now that we no longer distinguish between organisations and spaces in this copy.  There is no example of what kind of user would not be organisation level, and what this would mean.