Change table headers
The Billing manager has been swapped with the Org manager, so that the Org manager comes first as suggested.  This does make sense as the Org manager does have more permissions.

I shortened the suggested text "Spaces they're assigned to/they can access" to "Spaces assigned or can access" as the slash looked cluttered on a table header.  Additionally I think we're trying to avoid contractions these days as they are unclear to some people so removed the "they're".

I've held back on changing "name" to "username" as I would usually associate a username with the name of the person rather than their role, but would like other people's opinions on this.