Change space level role descriptions
As with the previous commit I've changed headers and the details lede for the sake of consistency.  However for the actual descriptions are again useful so I've left these.  I'm not sure if we want to use any of the suggested text in this case.

Suggested Text:

"v What can these roles do?
Space manager
A user who manages the space and its users

Space developer
A user who can deploy, run and manage apps, instantiate backing services and bind them to apps

Space auditor
A user who can view apps, users and resources used within the space (??? confirm)"