Fix travis
We need to compile the application on travis, to run the main tests.

The `main` tests need to wait for the application to be killed before
continuing with the tests, to make sure we get the correct thing tested.

We're upgrading the supertest library, which provides some additional
fixes for the cookie management from upstream. [1]

[1] DefinitelyTyped/DefinitelyTyped#27104