Scrap HMR and just use nodemon when watching build
We have been having constant issues with hot-module-reloading not
working or breaking whenever we make unrelated changes.

This gives up on the HMR approch and instead uses nodemon to watch the
dist dir and restart the server whenever it changes. This has a tiny
delay compared to HMR but it seems to be more reliable.

A nodemon webpack plugin is used to make the config simpler.