Allow authenticate UAA using user credentials
During the acceptance tests we will use user credentials of an
admin user to authenticate into UAA.

We implement this by adding a `authenticateUser` helper to
the `uaa` module, that would retrieve a new token using the
`password` method.

The retrieved token can be passed to a `UAAClient` constructor
omitting the client credentials. The existing implementation would
skip any authentication if a token is provided in the constructor.

In general this could be refactor to provide a more consistent
interface using the `UAAClient`, but we consider it out of scope
for the time being.

Signed-off-by: Richard Towers <richard.towers@digital.cabinet-office.gov.uk>