Add an acceptance-test using zombie+mocha
We want to test paas-admin integration with a real CloudFoundry
deployment in a end-to-end fashion.

We decided to use a web testing framework like zombie[1] for the
task, as it is simple has no dependencies to any JS engine that
we do not need.

The acceptance test expects some environment variables:

 - PAAS_ADMIN_BASE_URL: where the paas-admin is located.
 - CF_API_BASE_URL: api for CF
 - ADMIN_USERNAME and ADMIN_PASSWORD: creds for an admin user able
   to create orgs and users.

We implement some basic tests that will:

 - Prepare the test by creating a org manager and developer users
   and an organisation to test. We also add the corresponding
   tear down code.
 - Check that the user can login
 - Check that the user can see the org if it is a org manager.
 - Check that the user can invite the developer.

For the testing framework we use mocha[2]. We tried first to use
jest [2]  but it had some weird interaction with zombie
which we couldn't troubleshoot. Apparentally using jest, zombie
failed following redirects during the user authentication.

When using mocha, the process did not exit once the tests are
finished. We had to add `--exit` to force it.[3] We need to find
out why it did not exit properly.

[1] http://zombie.js.org/
[2] https://jestjs.io/
[3] https://mochajs.org/#--exit----no-exit

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>