Refactor the backing services listing
Now that we've made some recent changes to speed up admin tool, we've
forgotten that we require some extra fields for the backing service
listing.

These are now being added and the template is re-adjusted to fit the
purpose.

Test case has been added to prevent similar mistake happening again to
the same component.