Statements view get space names from  usage events
When displaying the statement for a given org we were pulling the
list of spaces for that org by doing a call to the CF API.

But this is problematic because:

 - BillingManagers do not have permissions to query the spaces, and
   we were preventing them to get the name of the space and filter
   by space because that.

 - The query to CF for spaces would return the current spaces. If a
   a spaced is deleted but still has events, it was not possible to
   filter based on that space.

But now paas-billing return the historic space name in the information
of the usage events [1], so we can rely on the events returned
by paas-billing to get the spaces.

In this commit we remove the call to the CF api, but use the info
in the events for the space guid and name to build the filter
dropbox and when displaying each stament row.

We do not add a specific test for this, but instead we remove the
"mocked" corresponding calls to the CF api to get the spaces.
Other unnecessary mocked calls have been removed too.

Note:

As collateral effect, now the filtering of spaces is based on the
historic data of the events:
 - Spaces with no events won't be displayed in the filtering dropbox.
 - Old space names will be used.
 - If a space is deleted, but there are events for it, it will be
   displayed in the filter dropbox.

We consider behaviour reasonable and within the user expectations.

[1] https://github.com/alphagov/paas-billing/pull/45

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>